Admin Header created and changes made to admin sidebar layout
