+++ frontend/src/App.js
@@ -1,7 +1,7 @@
 import {BrowserRouter as Router, Routes, Route} from 'react-router-dom'
 
 import Login from './pages/Login';
-import Admin from './components/layouts/Admin';
+import AdminLayout from './components/layouts/Admin';
 import FarmerLayout from './components/layouts/Farmer';
 
 function App() {
@@ -11,7 +11,7 @@ function App() {
         <div className='container'>
           <Routes>
               <Route path='/' element={<Login />} />
-              <Route path='/admin/*' element={<Admin />} />
+              <Route path='/admin/*' element={<AdminLayout />} />
               <Route path='/farmer/*' element={<FarmerLayout />} />
           </Routes>
         </div>
